FastAPI/requirements.txt
fastapi
pydantic
uvicorn[standard]
grpcio
grpcio-tools
